Refactor GameView layout calculation into dedicated helper
